+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Adamiszak 2025 fluconazole ([doi:10.3390/pharmaceutics17040488](https://doi.org/10.3390/pharmaceutics17040488)) -- hemato-oncologic pediatric patients aged 7 months to 18 years.
+- Add Wickramasinghe 2025 pamiparib ([doi:10.3390/pharmaceutics17040524](https://doi.org/10.3390/pharmaceutics17040524)) -- adults with newly diagnosed or recurrent glioblastoma.
 
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.